Move CertificationOfficer into domain/models
