+++ api/tests/integration/infrastructure/repositories/jury-session-repository_test.js
@@ -3,7 +3,7 @@ const { databaseBuilder, expect, catchErr } = require('../../../test-helper');
 const { NotFoundError } = require('../../../../lib/domain/errors');
 const JurySession = require('../../../../lib/domain/models/JurySession');
 const { statuses } = require('../../../../lib/domain/models/JurySession');
-const CertificationOfficer = require('../../../../lib/domain/read-models/CertificationOfficer');
+const CertificationOfficer = require('../../../../lib/domain/models/CertificationOfficer');
 const jurySessionRepository = require('../../../../lib/infrastructure/repositories/jury-session-repository');
 
 describe('Integration | Repository | JurySession', function() {
